requirements.txt
numpy
requests
tensorflow==1.15
soundfile
pyyaml
uwsgi
